+++ ckan/lib/search/query.py
@@ -2,16 +2,19 @@
 
 import re
 import logging
-
-from ckan.common import config
-import pysolr
-from paste.deploy.converters import asbool
-from paste.util.multidict import MultiDict
 import six
+import pysolr
 
-from ckan.lib.search.common import make_connection, SearchError, SearchQueryError
+from paste.deploy.converters import asbool
+from werkzeug.datastructures import MultiDict
+
 import ckan.logic as logic
 import ckan.model as model
+
+from ckan.common import config
+from ckan.lib.search.common import (
+    make_connection, SearchError, SearchQueryError
+)
 
 log = logging.getLogger(__name__)
 
@@ -20,7 +23,7 @@ _open_licenses = None
 VALID_SOLR_PARAMETERS = set([
     'q', 'fl', 'fq', 'rows', 'sort', 'start', 'wt', 'qf', 'bf', 'boost',
     'facet', 'facet.mincount', 'facet.limit', 'facet.field',
-    'extras', 'fq_list', 'tie', 'defType', 'mm'
+    'extras', 'fq_list', 'tie', 'defType', 'mm', 'df'
 ])
 
 # for (solr) package searches, this specifies the fields that are searched
@@ -211,7 +214,7 @@ class ResourceSearchQuery(SearchQuery):
             options.update(kwargs)
 
         context = {
-            'model':model,
+            'model': model,
             'session': model.Session,
             'search_query': True,
         }
@@ -219,6 +222,7 @@ class ResourceSearchQuery(SearchQuery):
         # Transform fields into structure required by the resource_search
         # action.
         query = []
+
         for field, terms in fields.items():
             if isinstance(terms, six.string_types):
                 terms = terms.split()
@@ -312,7 +316,9 @@ class PackageSearchQuery(SearchQuery):
             query['q'] = "*:*"
 
         # number of results
-        rows_to_return = min(1000, int(query.get('rows', 10)))
+        rows_to_return = int(query.get('rows', 10))
+        # query['rows'] should be a defaulted int, due to schema, but make
+        # certain, for legacy tests
         if rows_to_return > 0:
             # #1683 Work around problem of last result being out of order
             #       in SOLR 1.4