+++ ckan/lib/uploader.py
@@ -221,7 +221,7 @@ class ResourceUpload(object):
         upload_field_storage = resource.pop('upload', None)
         self.clear = resource.pop('clear_upload', None)
 
-        if config_mimetype_guess == 'file_ext':
+        if url and config_mimetype_guess == 'file_ext':
             self.mimetype = mimetypes.guess_type(url)[0]
 
         if isinstance(upload_field_storage, ALLOWED_UPLOAD_TYPES):