+++ ckanext/imageview/tests/test_view.py
@@ -9,22 +9,7 @@ from ckan.tests import helpers, factories
 
 
 class TestImageView(helpers.FunctionalTestBase):
-
-    @classmethod
-    def setup_class(cls):
-
-        super(TestImageView, cls).setup_class()
-
-        if not p.plugin_loaded('image_view'):
-            p.load('image_view')
-
-    @classmethod
-    def teardown_class(cls):
-        p.unload('image_view')
-
-        super(TestImageView, cls).teardown_class()
-
-        helpers.reset_db()
+    _load_plugins = ['image_view']
 
     @helpers.change_config('ckan.views.default_views', '')
     def test_view_shown_on_resource_page_with_image_url(self):